# Allow IntraL0 compaction in FIFO Compaction