Summary: Allow an option for users to do some compaction in FIFO compaction, to pay some write amplification for fewer number of files.

Test Plan: Add a unit test