@siying 2017-04-15
Some benchmark results:

Without the change:

[sdong@dev12075.prn1 /data/users/sdong/rocksdb7] TEST_TMPDIR=/dev/shm/ ./db_bench --benchmarks=readwhilewriting -compaction_style=2 --num=20000000 -level0_file_num_compaction_trigger=8 -bloom_bits=10 --threads=16

readwhilewriting :       0.725 micros/op 1378846 ops/sec;   54.7 MB/s (7070947 of 20000000 found)

With the change:

] TEST_TMPDIR=/dev/shm/ ./db_bench --benchmarks=readwhilewriting -compaction_style=2 --num=20000000 -level0_file_num_compaction_trigger=8 -bloom_bits=10 --threads=16 --fifo_compaction_allow_compaction --fifo_compaction_max_table_files_size_mb=1000

readwhilewriting :       0.485 micros/op 2062525 ops/sec;   78.0 MB/s (6766309 of 20000000 found)

With the change, the SST files under the directory after the compaction looks like this:

-rw-r--r--. 1 sdong users 241354289 Apr 14 17:45 000108.sst
-rw-r--r--. 1 sdong users 241391023 Apr 14 17:45 000125.sst
-rw-r--r--. 1 sdong users 241431557 Apr 14 17:45 000142.sst
-rw-r--r--. 1 sdong users  32490220 Apr 14 17:45 000144.sst
-rw-r--r--. 1 sdong users  32478382 Apr 14 17:45 000146.sst
-rw-r--r--. 1 sdong users  32498268 Apr 14 17:45 000148.sst