@ajkr 2017-04-24
just curious what was the motivation to introduce this for FIFO compaction?