@siying 2017-04-24
@ajkr a user sees high CPU overhead when there are many files in FIFO. I just want to create a solution for them to try.